Replace jitasm.h with AsmJit; enable x86 and x64 builds
Migrates the JIT callback emitter from the in-tree jitasm.h to the
maintained AsmJit library (vendored as a submodule). Both x86 cdecl and
x64 SystemV are now handled by a single source path; AsmJit's
FuncSignature + Compiler resolve the per-ABI register and stack layout.

CMake gains a BUILD_64BIT toggle that switches between -m32/-m64 builds
and produces metamod_i386.so / metamod_amd64.so accordingly. The arch
flag propagates globally so the AsmJit subbuild matches, and PIC is
forced on so the static AsmJit links into the shared object on x64.

Per-argument type information is now extracted at compile time alongside
the existing return-type extraction (getRetType<T>) so AsmJit can build
a correct FuncSignature for either ABI. Variadic engine functions are
not plugin-hookable in the AsmJit path on either ABI - those callbacks
emit a passthrough that calls the original directly. (No regression on
x64; minor regression on x86 vs the previous jitasm-emitted behavior.)

Mid-frame callback rebuild (meta_collect_fix_data / meta_apply_fix_data)
no longer reverse-engineers byte patterns from emitted code. CJit now
maintains a registry mapping post-call return addresses to stable
plugin-handler-slot identifiers, and esp_save stores the address of the
saved meta_globals backup directly. Pointer-size widening adopted in
mem_utils and metamod.cpp where the old uint32 casts assumed x86.

The 8968-line jitasm.h is removed; AsmJit ships at metamod/extern/asmjit
pinned to upstream master 0bd5787 (no release tags exist on AsmJit).

+++ metamod/src/mem_utils.h
@@ -19,7 +19,7 @@ public:
 	char *strdup(const char *string);
 	void deallocate_all();
 	size_t memory_used() const;
-	bool contain(uint32 addr);
+	bool contain(uintptr_t addr);
 	char *find_pattern(char *pattern, size_t len);
 
 	template<typename T>
@@ -45,5 +45,5 @@ private:
 
 bool  mem_compare(const char *addr, const char *pattern, size_t len);
 char *mem_find_pattern(char *pos, int range, const char *pattern, size_t len);
-char *mem_find_ref(char *pos, char *end, char opcode, uint32 ref, bool relative);
+char *mem_find_ref(char *pos, char *end, char opcode, uintptr_t ref, bool relative);
 char *mem_find_string_push(char *addr, const char *string, size_t len);
